devNotes-backend/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile
#
asgiref==3.7.2
    # via django
django==4.2.7
    # via
    #   -r requirements.in
    #   django-cors-headers
    #   djangorestframework
django-cors-headers==4.3.0
    # via -r requirements.in
djangorestframework==3.14.0
    # via -r requirements.in
pillow==10.1.0
    # via -r requirements.in
pytz==2023.3.post1
    # via djangorestframework
sqlparse==0.4.4
    # via django
